Adapts reading from stdin to node refactoring
